cleanup and re-organising file locations
